+++ js/bindings/bindingsTest.js
@@ -591,7 +591,6 @@ describe('Bindings', () => {
       const container = new b.Container({ name: 'foo', image: 'image' });
       const serv = new b.LoadBalancer({ name: 'foo', containers: [] });
       expect(container.getHostname()).to.equal('foo');
-      expect(serv.hostname()).to.equal('foo2');
       expect(serv.getHostname()).to.equal('foo2');
     });
     it('hostnames returned by uniqueHostname cannot be reused', () => {
@@ -799,7 +798,6 @@ describe('Bindings', () => {
     });
     it('get LoadBalancer hostname', () => {
       const foo = new b.LoadBalancer({ name: 'foo', containers: [] });
-      expect(foo.hostname()).to.equal('foo');
       expect(foo.getHostname()).to.equal('foo');
     });
   });